requirement.txt
ultralytics
albumentations
matplotlib
tqdm
Pillow
torch
